Fix workflow tab not showing for scheduled events
Fixes #1459.

The workflow tab in the event details of scheduled events was
not showing. It should now show again.
