+++ src/components/events/partials/modals/EventDetails.tsx
@@ -50,6 +50,7 @@ import ButtonLikeAnchor from "../../../shared/ButtonLikeAnchor";
 import { NOTIFICATION_CONTEXT } from "../../../../configs/modalConfig";
 import { unwrapResult } from "@reduxjs/toolkit";
 import { ParseKeys } from "i18next";
+import EventDetailsWorkflowSchedulingTab from "../ModalTabsAndPages/EventDetailsWorkflowSchedulingTab";
 
 export enum EventDetailsPage {
 	Metadata,
@@ -276,11 +277,10 @@ const EventDetails = ({
 						formikRef={formikRef}
 					/>
 				)}
-				{page === EventDetailsPage.Workflow &&
+				{page === EventDetailsPage.Workflow && !hasSchedulingProperties &&
 					((workflowTabHierarchy === "workflows" && (
 						<EventDetailsWorkflowTab
 							eventId={eventId}
-							formikRef={formikRef}
 						/>
 					)) ||
 						(workflowTabHierarchy === "workflow-details" && (
@@ -301,6 +301,12 @@ const EventDetails = ({
 								eventId={eventId}
 							/>
 						)))}
+				{page === EventDetailsPage.Workflow && hasSchedulingProperties &&
+					<EventDetailsWorkflowSchedulingTab
+						eventId={eventId}
+						formikRef={formikRef}
+					/>
+				}
 				{page === EventDetailsPage.AccessPolicy && (
 					<EventDetailsAccessPolicyTab
 						eventId={eventId}
